resolve validateToken issue, create AuthToken util function and AuthService class.
